+++ assets/javascripts/kitten/components/form/text-input-with-icon/index.js
@@ -6,6 +6,8 @@ import PropTypes from 'prop-types'
 import COLORS from '../../../constants/colors-config'
 import { VisuallyHidden } from '../../accessibility/visually-hidden'
 import classNames from 'classnames'
+import isEmpty from 'lodash/fp/isEmpty'
+import { ScreenConfig } from '../../../constants/screen-config'
 
 const StyledTextInputWithIcon = styled.div`
   position: relative;
@@ -49,6 +51,27 @@ const StyledTextInputWithIcon = styled.div`
       right: 0;
     }
   }
+
+  .k-Form-TextInputWithIcon__button {
+    border-radius: var(--input-button-border-radius);
+  }
+
+  &.k-Form-TextInput--orion {
+    --input-button-border-radius: ${pxToRem(4)};
+
+    &.k-Form-TextInput--big,
+    &.k-Form-TextInput--huge,
+    &.k-Form-TextInput--giant {
+      --input-button-border-radius: ${pxToRem(6)};
+
+      @media (min-width: ${ScreenConfig.M.min}px) {
+        --input-button-border-radius: ${pxToRem(8)};
+      }
+    }
+  }
+  &.k-Form-TextInput--rounded {
+    --input-button-border-radius: var(--input-height)
+  }
 `
 
 export const TextInputWithIcon = ({
@@ -57,6 +80,7 @@ export const TextInputWithIcon = ({
   iconPosition,
   accessibilityLabel,
   id,
+  buttonProps,
   ...others
 }) => {
   return (
@@ -81,14 +105,27 @@ export const TextInputWithIcon = ({
         )}
         disabled={disabled}
       />
-      <span
-        aria-hidden="true"
-        className={classNames('k-Form-TextInputWithIcon__icon', {
-          'k-Form-TextInputWithIcon__icon--disabled': disabled,
-        })}
-      >
-        {icon}
-      </span>
+      {isEmpty(buttonProps) ? (
+        <span
+          aria-hidden="true"
+          className={classNames('k-Form-TextInputWithIcon__icon', {
+            'k-Form-TextInputWithIcon__icon--disabled': disabled,
+          })}
+        >
+          {icon}
+        </span>
+      ) : (
+        <button
+          {...buttonProps}
+          className={classNames('k-u-reset-button',
+            'k-Form-TextInputWithIcon__button',
+            'k-Form-TextInputWithIcon__icon', {
+            'k-Form-TextInputWithIcon__icon--disabled': disabled,
+          })}
+        >
+          {icon}
+        </button>
+      )}
     </StyledTextInputWithIcon>
   )
 }
@@ -98,10 +135,12 @@ TextInputWithIcon.propTypes = {
   accessibilityLabel: PropTypes.string,
   icon: PropTypes.node.isRequired,
   iconPosition: PropTypes.oneOf(['left', 'right']),
+  buttonProps: PropTypes.object,
 }
 
 TextInputWithIcon.defaultProps = {
   accessibilityLabel: '',
   disabled: false,
   iconPosition: 'left',
+  buttonProps: {},
 }