+++ assets/javascripts/kitten/components/molecules/overlay/index.js
@@ -4,7 +4,6 @@ import styled from 'styled-components'
 import classNames from 'classnames'
 
 const StyledOverlay = styled.div`
-  position: absolute;
   left: 0;
   top: 0;
   right: 100%;
@@ -19,7 +18,14 @@ const StyledOverlay = styled.div`
   transition: visibility 0s ease, opacity 0.2s ease, bottom 0s, right 0s;
   transition-delay: 0.2s, 0s, 0.2s, 0.2s;
 
-  &.is-active {
+  &.k-Overlay--absolute {
+    position: absolute;
+  }
+  &.k-Overlay--fixed {
+    position: fixed;
+  }
+
+  &.k-Overlay--isActive {
     right: 0;
     bottom: 0;
 
@@ -61,12 +67,15 @@ export class Overlay extends Component {
   enableActiveState = () => this.setState({ isActive: true })
 
   render() {
-    const { className, zIndex, ...other } = this.props
+    const { className, zIndex, position, ...other } = this.props
     const { isActive } = this.state
     return (
       <StyledOverlay
         zIndex={zIndex}
-        className={classNames(isActive && 'is-active', className)}
+        className={classNames(className,
+          'k-Overlay', `k-Overlay--${position}`, {
+            'k-Overlay--isActive': isActive,
+          })}
         {...other}
       />
     )
@@ -79,6 +88,7 @@ Overlay.propTypes = {
   toggleEvent: PropTypes.string,
   closeEvent: PropTypes.string,
   openEvent: PropTypes.string,
+  position: PropTypes.oneOf('absolute', 'fixed')
 }
 
 Overlay.defaultProps = {
@@ -87,4 +97,5 @@ Overlay.defaultProps = {
   toggleEvent: '',
   closeEvent: '',
   openEvent: '',
+  position: 'absolute',
 }