+++ assets/javascripts/kitten/components/organisms/header-nav/components/floating-dropdown.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import classNames from 'classnames'
+import PropTypes from 'prop-types'
+
+export const FloatingDropdown = ({ children, className, visible }) => (
+  <div className={classNames('k-HeaderNav__floatingDropdown', className, {
+    'k-HeaderNav__floatingDropdown--visible': visible,
+  })}>{children}</div>
+)
+
+FloatingDropdown.propTypes = {
+  visible: PropTypes.bool,
+}
+
+FloatingDropdown.defaultProps = {
+  visible: false,
+}