+++ assets/javascripts/kitten/components/organisms/header-nav/stories/kisskissbankbank-new.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState, useEffect, useRef } from 'react'
 import HeaderNav from '../index'
 import {
   LendopolisLogo,
@@ -11,6 +11,9 @@ import {
   domElementHelper,
   useWindowWidth,
   useDeepCompareEffect,
+  Button,
+  LoaderWithParagraph,
+  Overlay,
 } from '../../../..'
 
 const HEADER_NAV_ID = 'kkbbAndCoHeaderNav'
@@ -83,6 +86,72 @@ const Navigation = () => (
   </>
 )
 
+const closeEvent = 'searchInput-overlay-close'
+const openEvent = 'searchInput-overlay-open'
+
+const SearchInput = () => {
+  const [isSearchMenuOpen, openSearchMenu] = useState(false)
+  const inputWrapperRef = useRef(null)
+
+  const handleSearchInputChange = (changeEvent) => {
+    openSearchMenu(changeEvent.target.value.length > 0)
+  }
+
+  useEffect(() => {
+    window.dispatchEvent(new Event(isSearchMenuOpen ? openEvent : closeEvent))
+    inputWrapperRef?.current.querySelector('#search-input').focus()
+
+    if (isSearchMenuOpen) {
+      inputWrapperRef.current.addEventListener('focusout', handleFocusOut)
+    } else {
+      inputWrapperRef.current.addEventListener('focusin', handleFocusIn)
+    }
+
+    return () => {
+      inputWrapperRef.current.removeEventListener('focusout', handleFocusOut)
+      inputWrapperRef.current.removeEventListener('focusin', handleFocusIn)
+    }
+  }, [isSearchMenuOpen])
+
+  const handleFocusOut = event => {
+    if (!inputWrapperRef.current || inputWrapperRef.current.contains(event.relatedTarget)) return
+
+    openSearchMenu(false)
+  }
+
+  const handleFocusIn = event => {
+    openSearchMenu(inputWrapperRef.current.querySelector('input').value.length > 0)
+  }
+
+  return (
+    <HeaderNav.SearchInput
+      inputProps={{
+        onChange: handleSearchInputChange,
+        id: 'search-input',
+      }}
+      expanded={isSearchMenuOpen}
+      buttonProps={{
+        type: 'submit',
+        'aria-label': 'Voir les résultats de la recherche'
+      }}
+      ref={inputWrapperRef}
+      openSearchMenu={openSearchMenu}
+    >
+      <HeaderNav.FloatingDropdown visible={isSearchMenuOpen}>
+        <LoaderWithParagraph
+          loaderPosition="top"
+          className="k-u-margin-bottom-double k-u-horizontal-auto"
+        >
+          Chargement
+        </LoaderWithParagraph>
+        <Button size="tiny" fluid variant="orion" type="submit">
+          Voir les résultats de la recherche
+        </Button>
+      </HeaderNav.FloatingDropdown>
+    </HeaderNav.SearchInput>
+  )
+}
+
 export const KissKissBankBankHeaderNavStoryNew = ({
   isLogged,
   isFixed,
@@ -101,101 +170,102 @@ export const KissKissBankBankHeaderNavStoryNew = ({
   }, [isLogged, windowWidth])
 
   return (
-    <HeaderNav
-      id={HEADER_NAV_ID}
-      isLogged={isLogged}
-      isFixed={isFixed}
-      quickAccessProps={{
-        href: '#mainContent',
-        text: 'Aller au contenu principal',
-        zIndex: 300,
-      }}
-      size={size}
-      borderStyle={borderStyle}
-    >
-      <HeaderNav.BurgerMenuNext dropdownContentWidth={burgerMenuWidth}>
-        <InnerBurgerMenu />
-      </HeaderNav.BurgerMenuNext>
+    <>
+      <Overlay
+        zIndex={1}
+        closeEvent={closeEvent}
+        openEvent={openEvent}
+        position="fixed"
+      />
 
-      <HeaderNav.Logo href="#">
-        <HeaderNav.Hidden min="xs">
-          <KissKissBankBankLogo height="25" />
+      <HeaderNav
+        id={HEADER_NAV_ID}
+        isLogged={isLogged}
+        isFixed={isFixed}
+        quickAccessProps={{
+          href: '#mainContent',
+          text: 'Aller au contenu principal',
+          zIndex: 300,
+        }}
+        size={size}
+        borderStyle={borderStyle}
+      >
+        <HeaderNav.BurgerMenuNext dropdownContentWidth={burgerMenuWidth}>
+          <InnerBurgerMenu />
+        </HeaderNav.BurgerMenuNext>
+
+        <HeaderNav.Logo href="#">
+          <HeaderNav.Hidden min="xs">
+            <KissKissBankBankLogo height="25" />
+          </HeaderNav.Hidden>
+
+          <HeaderNav.Hidden max="xxs">
+            <KissKissBankBankLogo />
+          </HeaderNav.Hidden>
+        </HeaderNav.Logo>
+
+        <HeaderNav.Hidden max="m">
+          <HeaderNav.Nav>
+            <Navigation />
+          </HeaderNav.Nav>
         </HeaderNav.Hidden>
 
-        <HeaderNav.Hidden max="xxs">
-          <KissKissBankBankLogo />
-        </HeaderNav.Hidden>
-      </HeaderNav.Logo>
+        <HeaderNav.Right padded>
+          <SearchInput />
 
-      <HeaderNav.Hidden max="m">
-        <HeaderNav.Nav>
-          <Navigation />
-        </HeaderNav.Nav>
-      </HeaderNav.Hidden>
-
-      <HeaderNav.Right>
-        <HeaderNav.Button
-          icon={<SearchIcon width="14" height="14" color={COLORS.font1} />}
-          backgroundColor={COLORS.background1}
-          backgroundColorHover={COLORS.line1}
-          color={COLORS.font1}
-          text="Rechercher"
-          href="#"
-          hiddenText={{ max: 'm' }}
-        />
-
-        <HeaderNav.Logged>
-          <HeaderNav.UserMenuNext
-            dropdownAnchorSide="right"
-            mobilePadding={false}
-          >
-            <HeaderNav.UserMenuNext.Button
-              hasArrow
-              backgroundColor={COLORS.background1}
-              backgroundColorHover={COLORS.line1}
-              backgroundColorActive={COLORS.line1}
-              color={COLORS.font1}
-              colorHover={COLORS.primary1}
-              colorActive={COLORS.font1}
+          <HeaderNav.Logged>
+            <HeaderNav.UserMenuNext
+              dropdownAnchorSide="right"
+              padding={false}
             >
-              <AvatarWithTextAndBadge className="k-u-margin-left-single@xs-down">
-                <AvatarWithTextAndBadge.Image src="/kitten.jpg">
-                  <AvatarWithTextAndBadge.Badge>2</AvatarWithTextAndBadge.Badge>
-                </AvatarWithTextAndBadge.Image>
+              <HeaderNav.UserMenuNext.Button
+                hasArrow
+                backgroundColor={COLORS.background1}
+                backgroundColorHover="transparent"
+                backgroundColorActive="transparent"
+                color={COLORS.font1}
+                colorHover={COLORS.primary1}
+                colorActive={COLORS.font1}
+              >
+                <AvatarWithTextAndBadge className="k-u-margin-left-single@xs-down">
+                  <AvatarWithTextAndBadge.Image src="/kitten.jpg">
+                    <AvatarWithTextAndBadge.Badge>2</AvatarWithTextAndBadge.Badge>
+                  </AvatarWithTextAndBadge.Image>
 
-                <AvatarWithTextAndBadge.Text
-                  className="k-u-hidden@xs-down"
-                  withEllipsisOverflow={true}
-                >
-                  <Text lineHeight="normal" weight="regular">
-                    Jean Charles Édouard
-                  </Text>
-                </AvatarWithTextAndBadge.Text>
-              </AvatarWithTextAndBadge>
-            </HeaderNav.UserMenuNext.Button>
+                  <AvatarWithTextAndBadge.Text
+                    className="k-u-hidden@xs-down"
+                    withEllipsisOverflow={true}
+                  >
+                    <Text lineHeight="normal" weight="regular">
+                      Jean Charles Édouard
+                    </Text>
+                  </AvatarWithTextAndBadge.Text>
+                </AvatarWithTextAndBadge>
+              </HeaderNav.UserMenuNext.Button>
 
-            <HeaderNav.UserMenuNext.Navigation>
-              <InnerUserMenu />
-            </HeaderNav.UserMenuNext.Navigation>
-          </HeaderNav.UserMenuNext>
-        </HeaderNav.Logged>
+              <HeaderNav.UserMenuNext.Navigation>
+                <InnerUserMenu />
+              </HeaderNav.UserMenuNext.Navigation>
+            </HeaderNav.UserMenuNext>
+          </HeaderNav.Logged>
 
-        <HeaderNav.LoggedOut>
-          <HeaderNav.Button
-            icon={
-              <HeaderNav.Hidden min="s">
-                <SearchIcon />
-              </HeaderNav.Hidden>
-            }
-            backgroundColor={COLORS.primary1}
-            backgroundColorHover={COLORS.primary2}
-            color={COLORS.background1}
-            text="Se connecter / S'inscrire"
-            href="#"
-            hiddenText={{ max: 'xs' }}
-          />
-        </HeaderNav.LoggedOut>
-      </HeaderNav.Right>
-    </HeaderNav>
+          <HeaderNav.LoggedOut>
+            <HeaderNav.Button
+              icon={
+                <HeaderNav.Hidden min="s">
+                  <SearchIcon />
+                </HeaderNav.Hidden>
+              }
+              backgroundColor={COLORS.primary1}
+              backgroundColorHover={COLORS.primary2}
+              color={COLORS.background1}
+              text="Se connecter / S'inscrire"
+              href="#"
+              hiddenText={{ max: 'xs' }}
+            />
+          </HeaderNav.LoggedOut>
+        </HeaderNav.Right>
+      </HeaderNav>
+    </>
   )
 }