+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `SearchIcon`: Update.
+- Feature: `TextInput`: Add `rounded` boolean prop, defaults to `false`.
+- Feature: `TextInputWithIcon`: Add `buttonProps` prop to use the icon as a button.
+- Feature: `Overlay`: Add `position` prop, defaults to `absolute`.
+
 ## [4.0.2] - 2021-07-27
 
 Fixes: